# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: untangle
    description: Untangle rumination sessions and logging
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /untangle/sessions:
    post:
      operationId: createSession
      tags: [untangle]
      summary: Create a new rumination session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
    get:
      operationId: listSessions
      tags: [untangle]
      summary: List all sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  /untangle/sessions/{id}:
    patch:
      operationId: updateSession
      tags: [untangle]
      summary: Update a session (e.g. mark timer completed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /untangle/chat:
    post:
      operationId: untangleChat
      tags: [untangle]
      summary: Send a message in a rumination conversation and get an AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UntangleChatBody"
      responses:
        "200":
          description: AI response with suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UntangleChatResponse"

  /untangle/transcribe:
    post:
      operationId: untangleTranscribe
      tags: [untangle]
      summary: Transcribe base64 audio to text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UntangleTranscribeBody"
      responses:
        "200":
          description: Transcribed text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UntangleTranscribeResponse"

  /untangle/ai-response:
    post:
      operationId: getAiResponse
      tags: [untangle]
      summary: Get an AI response for a rumination thought
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiResponseBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponseResult"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Session:
      type: object
      properties:
        id:
          type: integer
        ruminationThought:
          type: string
        aiResponse:
          type: string
        timerCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ruminationThought
        - timerCompleted
        - createdAt

    CreateSessionBody:
      type: object
      properties:
        ruminationThought:
          type: string
        aiResponse:
          type: string
      required:
        - ruminationThought

    UpdateSessionBody:
      type: object
      properties:
        timerCompleted:
          type: boolean
        aiResponse:
          type: string

    ChatHistoryItem:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
        - role
        - content

    UntangleChatBody:
      type: object
      properties:
        message:
          type: string
        mode:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryItem"
      required:
        - message
        - mode

    UntangleChatResponse:
      type: object
      properties:
        response:
          type: string
        isInsight:
          type: boolean
        suggestions:
          type: array
          items:
            type: string
      required:
        - response
        - isInsight
        - suggestions

    UntangleTranscribeBody:
      type: object
      properties:
        audio:
          type: string
        mimeType:
          type: string
      required:
        - audio
        - mimeType

    UntangleTranscribeResponse:
      type: object
      properties:
        text:
          type: string
      required:
        - text

    AiResponseBody:
      type: object
      properties:
        thought:
          type: string
      required:
        - thought

    AiResponseResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error